move tests from edge-test to their corresponding test files and remove edge-tests.ts
